feat(demo): add smoke app and tighten 32-bit build

+++ flipper/storage/filesystem_api_internal.h
@@ -19,7 +19,7 @@ typedef enum {
 struct File {
     FILE* file; /**< File sturct */
     DIR* dir;
-    uint32_t file_id; /**< File ID for internal references */
+    uintptr_t file_id; /**< File ID for internal references */
     FileType type;
     FS_Error error_id; /**< Standard API error from FS_Error enum */
     int32_t internal_error_id; /**< Internal API error value */
